@@ -1,6 +1,6 @@
 name: share_plus
 description: Flutter plugin for sharing content via the platform share UI, using the ACTION_SEND intent on Android and UIActivityViewController on iOS.
-version: 1.2.0
+version: 2.0.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
@@ -21,22 +21,19 @@ flutter:
 
 dependencies:
   meta: ^1.0.5
-  mime: ^0.9.7
+  mime: ^1.0.0-nullsafety
   flutter:
     sdk: flutter
-  share_plus_platform_interface: ^1.2.0
-  share_plus_linux: ^1.1.0
-  share_plus_windows: ^0.1.0
-  share_plus_web: ^0.1.0
+  share_plus_platform_interface: ^2.0.0-nullsafety
+  share_plus_linux: ^2.0.0-nullsafety
+  share_plus_windows: ^0.2.0-nullsafety
+  share_plus_web: ^0.2.0-nullsafety
 
 dev_dependencies:
-  test: ^1.3.0
-  mockito: ^3.0.0
   flutter_test:
     sdk: flutter
-  url_launcher: ^5.7.2
   pedantic: ^1.9.2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
