@@ -1,6 +1,6 @@
 name: share_plus_platform_interface
 description: A common platform interface for the share_plus plugin.
-version: 1.2.0
+version: 2.0.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
@@ -8,16 +8,16 @@ dependencies:
   flutter:
     sdk: flutter
   meta: ^1.1.8
-  mime: ^0.9.7
-  plugin_platform_interface: ^1.0.2
+  mime: ^1.0.0-nullsafety
+  plugin_platform_interface: ^1.1.0-nullsafety
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^3.0.0
+  mockito: ^5.0.0-nullsafety
   pedantic: ^1.9.2
   test: ^1.3.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
