@@ -2,10 +2,7 @@ name: share_plus_web
 description: Web platform implementation of share_plus
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
-# 0.1.y+z is compatible with 1.0.0, if you land a breaking change bump
-# the version to 2.0.0.
-# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
-version: 0.1.0
+version: 0.2.0-nullsafety.0
 
 flutter:
   plugin:
@@ -15,8 +12,8 @@ flutter:
         fileName: share_plus_web.dart
 
 dependencies:
-  share_plus_platform_interface: ^1.2.0
-  url_launcher: ^5.2.5
+  share_plus_platform_interface: ^2.0.0-nullsafety
+  url_launcher: ^6.0.0-nullsafety
   flutter:
     sdk: flutter
   flutter_web_plugins:
@@ -27,9 +24,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   pedantic: ^1.9.2
-  mockito: ^4.1.1
-  integration_test: ^1.0.2
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.20.0"
